Move ObjectAndMetadata
This will need to be referenced in more places like ChainDataProvider, so this move will make that simpler.
