+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/metadata/BlockAndMetaData.java
@@ -11,9 +11,8 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.api.blockselector;
+package tech.pegasys.teku.spec.datastructures.metadata;
 
-import tech.pegasys.teku.api.ObjectAndMetaData;
 import tech.pegasys.teku.spec.SpecMilestone;
 import tech.pegasys.teku.spec.datastructures.blocks.SignedBeaconBlock;
